Drop support for Python 2.7, 3.4, and 3.5.
Add support for Python 3.9 and 3.10.

Rename 'repoze.who._compat' to 'repoze.who._helpers': it no longer
provides any PY2 / PY3 straddling features.
