+++ repoze/who/plugins/redirector.py
@@ -1,14 +1,14 @@
+from urllib.parse import parse_qs
+from urllib.parse import urlencode
+from urllib.parse import urlparse
+from urllib.parse import urlunparse
 from webob.exc import HTTPFound
+
 from zope.interface import implementer
 
 from repoze.who.interfaces import IChallenger
-from repoze.who._compat import construct_url
-from repoze.who._compat import header_value
-from repoze.who._compat import parse_qs
-from repoze.who._compat import u
-from repoze.who._compat import urlencode
-from repoze.who._compat import urlparse
-from repoze.who._compat import urlunparse
+from repoze.who._helpers import construct_url
+from repoze.who._helpers import header_value
 
 @implementer(IChallenger)
 class RedirectorPlugin(object):
@@ -62,7 +62,7 @@ def make_plugin(login_url,
                 reason_param=None,
                 reason_header=None,
                ):
-    if login_url in (u(''), b'', None):
+    if login_url in (u'', b'', None):
         raise ValueError("No 'login_url'")
     if reason_header is not None and reason_param is None:
         raise Exception("Can't set 'reason_header' without 'reason_param'.")
