+++ repoze/who/restrict.py
@@ -1,8 +1,6 @@
 # Authorization middleware
 from pkg_resources import EntryPoint
 
-from repoze.who._compat import STRING_TYPES
-
 def authenticated_predicate():
     def _predicate(environ):
         return 'REMOTE_USER' in environ or 'repoze.who.identity' in environ
@@ -28,6 +26,6 @@ def make_authenticated_restriction(app, global_config, enabled=True):
 
 def make_predicate_restriction(app, global_config,
                                predicate, enabled=True, **kw):
-    if isinstance(predicate, STRING_TYPES):
+    if isinstance(predicate, str):
         predicate = EntryPoint.parse('x=%s' % predicate).resolve()
     return PredicateRestriction(app, predicate, enabled, **kw)
