+++ repoze/who/tests/test__auth_tkt.py
@@ -4,7 +4,7 @@ import unittest
 class AuthTicketTests(unittest.TestCase):
 
     def _getTargetClass(self):
-        from .._auth_tkt import AuthTicket
+        from repoze.who._auth_tkt import AuthTicket
         return AuthTicket
 
     def _makeOne(self, *args, **kw):
@@ -12,7 +12,7 @@ class AuthTicketTests(unittest.TestCase):
 
     def test_ctor_defaults(self):
         import hashlib
-        from .. import _auth_tkt
+        from repoze.who import _auth_tkt
         with _Monkey(_auth_tkt, time_mod=_Timemod):
             tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4')
         self.assertEqual(tkt.secret, 'SEEKRIT')
@@ -94,7 +94,7 @@ class AuthTicketTests(unittest.TestCase):
         self.assertEqual(tkt.digest_algo, hashlib.sha1)
 
     def test_digest(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4', tokens=('a', 'b'),
                             user_data='DATA', time=_WHEN,
                             cookie_name='oatmeal', secure=True)
@@ -103,7 +103,7 @@ class AuthTicketTests(unittest.TestCase):
         self.assertEqual(tkt.digest(), digest)
 
     def test_cookie_value_wo_tokens_or_userdata(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4', time=_WHEN)
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
                                   '', '', hashlib.md5)
@@ -111,7 +111,7 @@ class AuthTicketTests(unittest.TestCase):
                          '%s%08xUSERID!' % (digest, _WHEN))
 
     def test_cookie_value_w_tokens_and_userdata(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4', tokens=('a', 'b'),
                             user_data='DATA', time=_WHEN)
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
@@ -120,8 +120,8 @@ class AuthTicketTests(unittest.TestCase):
                          '%s%08xUSERID!a,b!DATA' % (digest, _WHEN))
 
     def test_cookie_not_secure_wo_tokens_or_userdata(self):
-        from .._auth_tkt import calculate_digest, hashlib
-        from .._compat import encodestring
+        from repoze.who._auth_tkt import calculate_digest, hashlib
+        from repoze.who._helpers import encodestring
         tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4', time=_WHEN,
                             cookie_name='oatmeal')
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
@@ -134,8 +134,8 @@ class AuthTicketTests(unittest.TestCase):
         self.assertEqual(cookie['oatmeal']['secure'], '')
 
     def test_cookie_secure_w_tokens_and_userdata(self):
-        from .._auth_tkt import calculate_digest, hashlib
-        from .._compat import encodestring
+        from repoze.who._auth_tkt import calculate_digest, hashlib
+        from repoze.who._helpers import encodestring
         tkt = self._makeOne('SEEKRIT', 'USERID', '1.2.3.4', tokens=('a', 'b'),
                             user_data='DATA', time=_WHEN,
                             cookie_name='oatmeal', secure=True)
@@ -152,7 +152,7 @@ class AuthTicketTests(unittest.TestCase):
 class BadTicketTests(unittest.TestCase):
 
     def _getTargetClass(self):
-        from .._auth_tkt import BadTicket
+        from repoze.who._auth_tkt import BadTicket
         return BadTicket
 
     def _makeOne(self, *args, **kw):
@@ -173,11 +173,11 @@ class Test_parse_ticket(unittest.TestCase):
 
     def _callFUT(self, secret='SEEKRIT', ticket=None,
                  ip='1.2.3.4', digest="md5"):
-        from .._auth_tkt import parse_ticket
+        from repoze.who._auth_tkt import parse_ticket
         return parse_ticket(secret, ticket, ip, digest)
 
     def test_bad_timestamp(self):
-        from .._auth_tkt import BadTicket
+        from repoze.who._auth_tkt import BadTicket
         TICKET = '12345678901234567890123456789012XXXXXXXXuserid!'
         try:
             self._callFUT(ticket=TICKET)
@@ -188,7 +188,7 @@ class Test_parse_ticket(unittest.TestCase):
             self.fail('Did not raise')
 
     def test_no_bang_after_userid(self):
-        from .._auth_tkt import BadTicket
+        from repoze.who._auth_tkt import BadTicket
         TICKET = '1234567890123456789012345678901201020304userid'
         try:
             self._callFUT(ticket=TICKET)
@@ -198,7 +198,7 @@ class Test_parse_ticket(unittest.TestCase):
             self.fail('Did not raise')
 
     def test_wo_tokens_or_data_bad_digest(self):
-        from .._auth_tkt import BadTicket
+        from repoze.who._auth_tkt import BadTicket
         TICKET = '1234567890123456789012345678901201020304userid!'
         try:
             self._callFUT(ticket=TICKET)
@@ -208,7 +208,7 @@ class Test_parse_ticket(unittest.TestCase):
             self.fail('Did not raise')
 
     def test_wo_tokens_or_data_ok_digest(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
                                   '', '', hashlib.md5)
         TICKET = '%s%08xUSERID!' % (digest, _WHEN)
@@ -219,7 +219,7 @@ class Test_parse_ticket(unittest.TestCase):
         self.assertEqual(user_data, '')
 
     def test_w_tokens_and_data_ok_digest(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
                                   'a,b', 'DATA', hashlib.md5)
         TICKET = '%s%08xUSERID!a,b!DATA' % (digest, _WHEN)
@@ -230,7 +230,7 @@ class Test_parse_ticket(unittest.TestCase):
         self.assertEqual(user_data, 'DATA')
 
     def test_w_tokens_and_data_ok_alternate_digest(self):
-        from .._auth_tkt import calculate_digest, hashlib
+        from repoze.who._auth_tkt import calculate_digest, hashlib
         digest = calculate_digest('1.2.3.4', _WHEN, 'SEEKRIT', 'USERID',
                                   'a,b', 'DATA', hashlib.sha256)
         TICKET = '%s%08xUSERID!a,b!DATA' % (digest, _WHEN)
@@ -248,29 +248,28 @@ class Test_helpers(unittest.TestCase):
 
     def test_ints_to_bytes(self):
         from struct import pack
-        from .._auth_tkt import ints2bytes
+        from repoze.who._auth_tkt import ints2bytes
         self.assertEqual(ints2bytes([1, 2, 3, 4]), pack('>BBBB', 1, 2, 3, 4))
         
     def test_encode_ip_timestamp(self):
         from struct import pack
-        from .._auth_tkt import encode_ip_timestamp
+        from repoze.who._auth_tkt import encode_ip_timestamp
         self.assertEqual(encode_ip_timestamp('1.2.3.4', _WHEN),
                          pack('>BBBBL', 1, 2, 3, 4, _WHEN))
 
     def test_maybe_encode_bytes(self):
-        from .._auth_tkt import maybe_encode
+        from repoze.who._auth_tkt import maybe_encode
         foo = b'foo'
         self.assertTrue(maybe_encode(foo) is foo)
 
     def test_maybe_encode_native_string(self):
-        from .._auth_tkt import maybe_encode
+        from repoze.who._auth_tkt import maybe_encode
         foo = 'foo'
         self.assertEqual(maybe_encode(foo), b'foo')
 
     def test_maybe_encode_unicode(self):
-        from .._auth_tkt import maybe_encode
-        from .._compat import u
-        foo = u('foo')
+        from repoze.who._auth_tkt import maybe_encode
+        foo = u'foo'
         self.assertEqual(maybe_encode(foo), b'foo')
