+++ repoze/who/tests/test_config.py
@@ -40,7 +40,7 @@ class TestWhoConfig(unittest.TestCase):
         self.assertEqual(len(config.mdproviders), 0)
 
     def test_parse_empty_file(self):
-        from repoze.who._compat import StringIO
+        from io import StringIO
         config = self._makeOne()
         config.parse(StringIO())
         self.assertEqual(config.request_classifier, None)
