+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist = 
-    py27,pypy,py34,py35,py36,py37,py38,pypy3,cover,docs
+    py36,py37,py38,py39,py310,pypy3,cover,docs
 
 [testenv]
 commands = 
@@ -18,7 +18,7 @@ setenv =
 [testenv:cover]
 skip_install = true
 basepython =
-    python3.8
+    python3.10
 commands = 
     coverage combine
     coverage report --fail-under=100 --show-missing --omit="*fixture*"
@@ -30,7 +30,7 @@ setenv =
 
 [testenv:docs]
 basepython =
-    python3.8
+    python3.10
 commands = 
     sphinx-build -b html -d docs/_build/doctrees docs docs/_build/html
     sphinx-build -b doctest -d docs/_build/doctrees docs docs/_build/doctest
